+++ ee/tables/wmitable/wmitable.go
@@ -13,7 +13,7 @@ import (
 	"github.com/kolide/launcher/ee/dataflatten"
 	"github.com/kolide/launcher/ee/tables/dataflattentable"
 	"github.com/kolide/launcher/ee/tables/tablehelpers"
-	"github.com/kolide/launcher/pkg/wmi"
+	"github.com/kolide/launcher/ee/wmi"
 
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
