+++ ee/wmi/wmi.go
