+++ ee/wmi/wmi_test.go
